requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
whitenoise==5.0.1
idna==2.9
oauthlib==3.1.0
pip-lock==2.1.0
pipenv==2018.11.26
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4

